refactor(cli): introduce `titus enum` parent command and dedup scan pipeline
Group the 8 remote-scanning commands (github, gitlab, slack, notion, linear,
confluence, jira, sharepoint) under a new `titus enum` parent, declutter
`titus --help`, and collapse the copy-pasted ~55-line scan pipeline into a
single `runEnumScan` helper (net -977 lines).

- Hoist shared flags (--output/--format/--rules/--rules-include/
  --rules-exclude/--ruleset/--include-noisy) to `enum` persistent flags.
- Add `enum microsoft` parent holding Graph auth flags (--client-id/
  --tenant-id/--token/--refresh-token); `sharepoint` is its first subcommand.
- Keep all 8 old top-level commands as hidden, deprecated aliases so existing
  scripts keep working (github/gitlab retain their `scan` subcommand).
- github/gitlab keep all clone logic; `scan` (local), `serve`, `rules`,
  `report`, `explore`, `version` are unchanged and stay top-level.

Intended behavior changes:
- notion now honors --rules*/--include-noisy (was hardcoded to ignore them);
  its scan-complete status lines move stderr -> stdout (uniform with others).
- gitlab matcher ContextLines normalizes 0 -> 3.
- --format json no longer corrupted by human summary lines in github/gitlab
  (the json branch returns before the summary prints).

Verification: CGO=0 and CGO=1 (vectorscan) builds pass; go vet clean;
go test ./cmd/titus/... = 198 pass (incl. -race). Local `titus scan --git`
benchmark unchanged vs main: kubernetes 409=409 findings, gitlab 71887=71887
findings, timing within noise (scan engine code is byte-identical).
